refactor: add folder test in respository
