+++ src/@seedwork/domain/repository/__tests__/in-memory.repository.spec.ts
@@ -1,6 +1,6 @@
-import { NotFoundError } from "../../errors/in-memory/not-found.error"
-import { Entity } from "../entity/entity"
-import { InMemoryRepository } from "./in-memory.repository"
+import { NotFoundError } from "../../../errors/in-memory/not-found.error"
+import { Entity } from "../../entity/entity"
+import { InMemoryRepository } from "../in-memory.repository"
 
 
 type StubEntityProps = {
